Added an atmosphere shader
