+++ src/shaders/atmosphere/atmosphereVertex.glsl
@@ -1,9 +1,7 @@
-varying vec2 vUv;
 varying vec3 vPosition;
 varying vec3 vNormal;
 
 void main() {
-    vUv = uv;
     vPosition = (modelMatrix * vec4(position, 1.0)).xyz;
     vNormal = (modelMatrix * vec4(normal, 0.0)).xyz;
